+++ src/lib/Client.js
@@ -17,10 +17,9 @@ function mixin(obj1, obj2) {
 };
 
 
-function Client(uri, arrayBuffer) {
+function Client(uri) {
 	EventEmitter.call(this);
 	this.uri = uri;
-	this.arrayBuffer = arrayBuffer;
 	this.ws = undefined;
 	this.serverTimeOffset = 0;
 	this.user = undefined;
@@ -85,7 +84,7 @@ Client.prototype.connect = function() {
 		// browseroni
 		this.ws = new WebSocket(this.uri);
 	}
-	if (this.arrayBuffer) this.ws.binaryType = "arraybuffer";
+	this.ws.binaryType = "arraybuffer";
 	this.emit("ws created");
 	var self = this;
 	this.ws.addEventListener("close", function(evt) {
