+++ src/main.js
@@ -1,3 +1,4 @@
+require('./util');
 global.exitHook = require('async-exit-hook');
 global.Discord = require('discord.js');
 global.fs = require('fs');
@@ -41,11 +42,11 @@ process.on('uncaughtException', error => {
 dClient.once('ready', () => {
 	console.log('Discord Client Ready');
 
-	require('./commands.js');
-		require('./colorroles.js');
-		require('./mppbridger.js');
-	require('./screenshotter.js');
-	require('./misc.js');
+	require('./commands');
+		require('./colorroles');
+		require('./mppbridger');
+	require('./screenshotter');
+	require('./misc');
 
 });
 dClient.on('error', console.error);
