+++ src/util.js
@@ -1,0 +1,3 @@
+Array.prototype.random = function () {
+    return this[Math.floor(Math.random() * this.length)]
+}
